test(web-app): add auth options test
