+++ docker/web-app/tsconfig.test.json
@@ -25,7 +25,8 @@
     "src/components/__tests__/uiUxTasks.test.ts",
     "src/lib/server/__tests__/settingsDB.test.ts",
     "src/app/api/policy/__tests__/evaluate.test.ts",
-    "src/types/externals.d.ts"
+    "src/types/externals.d.ts",
+    "src/lib/auth.test.ts"
   ],
   "exclude": []
 }
